Allow diagnostics download to default to the latest created bundle
