+++ pkg/cmd/diagnostics/diagnostics_download.go
@@ -17,12 +17,21 @@ func newDiagnosticsDownloadCommand(ctx api.Context) *cobra.Command {
 		// the <bundle-id> is used to indicate what the expected positional arg is in the help output
 		Use:   "download <bundle-id>",
 		Short: "Download diagnostics bundle",
-		Args:  cobra.ExactArgs(1),
+		Args:  cobra.MaximumNArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			c := pluginutil.HTTPClient("")
 			client := diagnostics.NewClient(c)
 
-			id := args[0]
+			var id string
+			if len(args) == 0 {
+				bundle, err := latestBundle(client)
+				if err != nil {
+					return err
+				}
+				id = bundle.ID
+			} else {
+				id = args[0]
+			}
 
 			if outputPath == "" {
 				outputPath = fmt.Sprintf("%s.zip", id)
